fix merge conflicts

+++ gadgets/src/bits/boolean_input.rs
@@ -138,7 +138,7 @@ impl<F: PrimeField, CF: PrimeField> AllocGadget<Vec<F>, CF> for BooleanInputGadg
 
         // Step 3: allocate the CF field elements as input
         let mut src_booleans = Vec::<Boolean>::new();
-        for (i, chunk) in src_bits.chunks(capacity).enumerate() {
+        for (i, chunk) in src_bits.chunks(capacity as usize).enumerate() {
             let elem = CF::from_repr(<CF as PrimeField>::BigInteger::from_bits_le(chunk)).unwrap();
 
             let elem_gadget = FpGadget::<CF>::alloc_input(cs.ns(|| format!("alloc_elem_{}", i)), || Ok(elem))?;
